Fix ActionMenu + tests
