+++ mobile/src/ActionMenu/styles.js
@@ -5,7 +5,6 @@ import { colors } from 'app/settings';
 
 export default StyleSheet.create({
   Fab: {
-    color: colors.textAltPrimaryColor,
     backgroundColor: colors.secondaryColor,
   },
 });
